requirements.txt
PyYAML
paramiko
autopep8
